# Vectorize bucketing